+++ src/ksm/endpoints/endpoints.go
@@ -2,7 +2,10 @@ package endpoints
 
 import (
 	"fmt"
+	"net/http"
 	"net/url"
+	"path/filepath"
+	"time"
 
 	"net"
 
@@ -10,6 +13,7 @@ import (
 
 	"github.com/newrelic/infra-integrations-beta/integrations/kubernetes/src/endpoints"
 	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
 )
 
 const (
@@ -20,31 +24,75 @@ const (
 	ksmQualifiedName = "kube-state-metrics.kube-system.svc.cluster.local"
 	ksmDNSService    = "http-metrics"
 	ksmDNSProto      = "tcp"
+	acceptHeader     = `application/vnd.google.protobuf;proto=io.prometheus.client.MetricFamily;encoding=delimited;q=0.7,text/plain;version=0.0.4;q=0.3`
 )
 
 // ksmDiscoverer implements Discoverer interface by using official Kubernetes' Go client
 type ksmDiscoverer struct {
 	lookupSRV func(service, proto, name string) (cname string, addrs []*net.SRV, err error)
-	client    endpoints.KubernetesClient
+	apiClient endpoints.KubernetesClient
+	logger    *logrus.Logger
 }
 
-// Discover returns the schema://host:port URL part of Kube State Metrics
-func (sd ksmDiscoverer) Discover() (url.URL, error) {
-	var endpoint url.URL
+// ksm implements Client interface
+type ksm struct {
+	httpClient *http.Client
+	endpoint   url.URL
+	nodeIP     string
+	logger     *logrus.Logger
+}
 
+// TODO: cache it
+func (sd *ksmDiscoverer) Discover(timeout time.Duration) (endpoints.Client, error) {
+
+	var endpoint url.URL
 	endpoint, err := sd.dnsDiscover()
 	if err != nil {
 		// if DNS discovery fails, we dig into Kubernetes API to get the service data
 		endpoint, err = sd.apiDiscover()
+		if err != nil {
+			return nil, fmt.Errorf("failed to discover kube-state-metrics endpoint, got error: %s", err)
+		}
 	}
 
 	// KSM and Prometheus only work with HTTP
 	endpoint.Scheme = "http"
-	return endpoint, err
+	nodeIP, err := sd.nodeIP()
+	if err != nil {
+		return nil, fmt.Errorf("failed to discover nodeIP with kube-state-metrics, got error: %s", err)
+	}
+
+	return &ksm{
+		nodeIP:   nodeIP,
+		endpoint: endpoint,
+		httpClient: &http.Client{
+			Timeout: timeout,
+		},
+		logger: sd.logger,
+	}, nil
+}
+
+func (c *ksm) NodeIP() string {
+	return c.nodeIP
+}
+
+func (c *ksm) Do(method, path string) (*http.Response, error) {
+	e := c.endpoint
+	e.Path = filepath.Join(c.endpoint.Path, path)
+
+	r, err := http.NewRequest(method, e.String(), nil)
+	if err != nil {
+		return nil, fmt.Errorf("Error creating %s request to: %s. Got error: %s ", method, e.String(), err)
+	}
+	r.Header.Set("Accept", acceptHeader)
+
+	c.logger.Debugf("Calling kube-state-metrics endpoint: %s", r.URL.String())
+
+	return c.httpClient.Do(r)
 }
 
 // dnsDiscover uses DNS to discover KSM
-func (sd ksmDiscoverer) dnsDiscover() (url.URL, error) {
+func (sd *ksmDiscoverer) dnsDiscover() (url.URL, error) {
 	var endpoint url.URL
 	_, addrs, err := sd.lookupSRV(ksmDNSService, ksmDNSProto, ksmQualifiedName)
 	if err == nil {
@@ -57,10 +105,10 @@ func (sd ksmDiscoverer) dnsDiscover() (url.URL, error) {
 }
 
 // apiDiscover uses Kubernetes API to discover KSM
-func (sd ksmDiscoverer) apiDiscover() (url.URL, error) {
+func (sd *ksmDiscoverer) apiDiscover() (url.URL, error) {
 	var endpoint url.URL
 
-	services, err := sd.client.FindServiceByLabel(ksmAppLabelName, ksmAppLabelValue)
+	services, err := sd.apiClient.FindServiceByLabel(ksmAppLabelName, ksmAppLabelValue)
 	if err != nil {
 		return endpoint, err
 	}
@@ -91,8 +139,8 @@ func (sd ksmDiscoverer) apiDiscover() (url.URL, error) {
 	return endpoint, fmt.Errorf("could not guess the Kube State Metrics host/port")
 }
 
-func (sd ksmDiscoverer) NodeIP() (string, error) {
-	pods, err := sd.client.FindPodsByLabel(ksmAppLabelName, ksmAppLabelValue)
+func (sd *ksmDiscoverer) nodeIP() (string, error) {
+	pods, err := sd.apiClient.FindPodsByLabel(ksmAppLabelName, ksmAppLabelValue)
 	if err != nil {
 		return "", err
 	}
@@ -101,29 +149,28 @@ func (sd ksmDiscoverer) NodeIP() (string, error) {
 	}
 	// In case there are multiple pods for the same service, we must be sure we always show the Node IP of the
 	// same pod. So we chose, for example, the HostIp with highest precedence in alphabetical order
-	var hostIP string
+	var nodeIP string
 	for _, pod := range pods.Items {
-		if pod.Status.HostIP != "" && (hostIP == "" || strings.Compare(pod.Status.HostIP, hostIP) < 0) {
-			hostIP = pod.Status.HostIP
+		if pod.Status.HostIP != "" && (nodeIP == "" || strings.Compare(pod.Status.HostIP, nodeIP) < 0) {
+			nodeIP = pod.Status.HostIP
 		}
 	}
-	if hostIP == "" {
+	if nodeIP == "" {
 		return "", errors.New("no HostIP address found for KSM node")
 	}
-	return hostIP, nil
+	return nodeIP, nil
 }
 
 // NewKSMDiscoverer instantiates a new Discoverer
-func NewKSMDiscoverer() (endpoints.Discoverer, error) {
+func NewKSMDiscoverer(logger *logrus.Logger) (endpoints.Discoverer, error) {
 	var discoverer ksmDiscoverer
 	var err error
 
-	discoverer.client, err = endpoints.NewKubernetesClient()
+	discoverer.apiClient, err = endpoints.NewKubernetesClient()
 	if err != nil {
 		return nil, err
 	}
-
 	discoverer.lookupSRV = net.LookupSRV
-
-	return discoverer, nil
+	discoverer.logger = logger
+	return &discoverer, nil
 }